aws-deploy/requirements.txt
boto3==1.16.59
botocore==1.19.59
jmespath==0.10.0
mysql-connector-python==8.0.23
protobuf==3.14.0
python-dateutil==2.8.1
python-dotenv==0.15.0
s3transfer==0.3.4
six==1.15.0
urllib3==1.26.2
